requirements.txt
snowflake-connector-python
snowflake-connector-python[pandas]
beautifulsoup4
requests